Fragments
Layouts Categories  - Changes:
Several changes in Layouts to set the fragments in better ways.

Fragments Gastronomy - Created:
Fragments for the menu into categories, restaurants and typical meals was created and given the proper design.

Fragments Code - In Process:
The buttons was coded to bring the new fragments but already don't works.