All Activities with Fragments
Tourism Activity Options - Edited:
I did change the name to manage it as only a window to select the option in the category, and I coded to send the properly name Fragment to display in Tourism Activity.

Tourism Activity - Created:
Created as the activity to charge the fragments fill of data, created the function to receive the proper name fragment and coded the slide bar functions.

Tourism Fragments - Created

Variable Names - Edited:
Many variable names was edited both in layouts as in the activities to work with them properly.
Also many imports was wrong called when the code was copy and pasted, maybe there is some more in other activities. BE CAREFULLY WITH IT AND CORRECT THEM EARLIER POSIBLE.